+++ src/api/bff/enrollment.ts
@@ -1,6 +1,6 @@
 import qs from 'query-string';
-import {client} from './client';
-import {stringifyUrl} from './utils';
+import {client} from '../client';
+import {stringifyUrl} from '../utils';
 
 type EnrollmentResponse =
   | {
