+++ src/api/bff/geocoder.ts
@@ -1,8 +1,8 @@
 import {FOCUS_LATITUDE, FOCUS_LONGITUDE, TARIFF_ZONE_AUTHORITY} from '@env';
-import {Feature, Coordinates} from '../sdk';
-import {client} from './client';
+import {Feature, Coordinates} from '../../sdk';
+import {client} from '../client';
 import qs from 'query-string';
-import {stringifyUrl} from './utils';
+import {stringifyUrl} from '../utils';
 import {AxiosRequestConfig} from 'axios';
 
 export const FOCUS_ORIGIN: Coordinates = {
