+++ src/api/bff/trips.ts
@@ -1,4 +1,4 @@
-import {client} from './client';
+import {client} from '../client';
 import {AxiosRequestConfig} from 'axios';
 import {
   type BookingAvailabilityQueryVariables,
