+++ src/modules/map/ExploreLocationMap.tsx
@@ -2,7 +2,7 @@ import {
   getCurrentCoordinatesGlobal,
   useGeolocationContext,
 } from '@atb/modules/geolocation';
-import {FOCUS_ORIGIN} from '@atb/api/geocoder';
+import {FOCUS_ORIGIN} from '@atb/api/bff/geocoder';
 import {StyleSheet} from '@atb/theme';
 import MapboxGL, {LocationPuck} from '@rnmapbox/maps';
 import {Feature} from 'geojson';
