+++ src/screen-components/travel-details-screens/use-current-trip-pattern-with-updates.ts
@@ -1,7 +1,7 @@
 import {TripPattern} from '@atb/api/types/trips';
 import {AxiosError} from 'axios';
 import {useCallback, useEffect, useState} from 'react';
-import {singleTripSearch} from '@atb/api/trips';
+import {singleTripSearch} from '@atb/api/bff/trips';
 import {usePollableResource} from '@atb/utils/use-pollable-resource';
 
 type TripPatternUpdate = {
