+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/Dashboard_RootScreen/components/DeparturesWidget.tsx
@@ -1,5 +1,5 @@
 import {Add, Edit} from '@atb/assets/svg/mono-icons/actions';
-import {StopPlaceInfo} from '@atb/api/departures/types';
+import {StopPlaceInfo} from '@atb/api/bff/types';
 import {useBottomSheetContext} from '@atb/components/bottom-sheet';
 import {Button} from '@atb/components/button';
 import {ThemeText} from '@atb/components/text';
